refactor empty state
