+++ components/Search/index.js
@@ -5,6 +5,7 @@ import Form from './Form'
 import Filters from './Filters'
 import Sort from './Sort'
 import Results from './Results'
+import CheatSheet from './CheatSheet'
 
 import { Center, mediaQueries } from '@project-r/styleguide'
 import withSearchRouter from './withSearchRouter'
@@ -20,7 +21,7 @@ const styles = {
   })
 }
 
-export default withSearchRouter(({ cleanupUrl }) => {
+export default withSearchRouter(({ cleanupUrl, empty }) => {
   useEffect(() => {
     cleanupUrl()
   }, [])
@@ -28,9 +29,15 @@ export default withSearchRouter(({ cleanupUrl }) => {
   return (
     <Center {...styles.container}>
       <Form />
-      <Filters />
-      <Sort />
-      <Results />
+      {empty ? (
+        <CheatSheet />
+      ) : (
+        <>
+          <Filters />
+          <Sort />
+          <Results />
+        </>
+      )}
     </Center>
   )
 })
